Shift wiki specific methods out of the Skill
