+++ wiki/pymediawiki.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import re
 from urllib3.exceptions import HTTPError
 
 from mediawiki import (
@@ -34,8 +35,9 @@ EXCLUDED_IMAGES = [
 class Wiki():
     """Interface to Wikipedia using pymediawiki."""
 
-    def __init__(self, lang) -> None:
+    def __init__(self, lang, auto_more) -> None:
         self.default_lang = lang
+        self.auto_more = auto_more
         try:
             self.wiki = MediaWiki(lang=lang)
         except HTTPError:
@@ -81,6 +83,26 @@ class Wiki():
 
         return image or DEFAULT_IMAGE
 
+    @staticmethod
+    def get_disambiguation_page(results: list([str])) -> str:
+        """Get the disambiguation page title from a set of results.
+
+        Note that some disambiguation pages aren't explicitly labelled as one.
+        The only guaranteed way to know is to fetch the page and catch a
+        DisambiguationError eg "George Church"
+
+        Args:
+            results: list of wikipedia pages
+        Returns:
+            disambiguation page title or None
+        """
+        try:
+            page_title = next(
+                page for page in results if "(disambiguation)" in page)
+        except StopIteration:
+            page_title = None
+        return page_title
+
     def get_page(self, title: str) -> MediaWikiPage:
         """Get the specified wiki page."""
         try:
@@ -97,6 +119,42 @@ class Wiki():
         self.set_language('en')
         random_page = self.wiki.random(pages=1)
         return self.get_page(random_page)
+
+    def get_summary_intro(self, page: MediaWikiPage) -> tuple([str, int]):
+        """Get a short summary of the page.
+
+        About auto_more (bool): default False
+            Set by cq_auto_more attribute in mycroft.conf
+            If true will read 20 sentences of abstract for any query.
+            If false will read first 2 sentences and wait for request to read more.
+
+        Args:
+            page: wiki page containing a summary
+        Returns:
+            trimmed answer, number of sentences
+        """
+        length = 20 if self.auto_more else 2
+        answer = self.summarize_page(page, sentences=length)
+        if not self.auto_more and len(answer) > 250:
+            answer = self.summarize_page(page, sentences=1)
+        return answer, length
+
+    def get_summary_next_lines(self, page: MediaWikiPage, previous_lines: int, num_lines: int = 5) -> tuple([str, int]):
+        """Get the next summary lines to be read.
+
+        Args:
+            page: wiki page containing a summary
+            previous_lines: number of sentences already read
+            num_lines: number of new lines to return
+        Returns:
+            next lines of summary,
+            total length of summary read so far ie previous_lines + num_lines
+        """
+        total_summary_read = previous_lines + num_lines
+        previously_read = page.summarize(sentences=previous_lines)
+        next_summary_section = self.summarize_page(
+            page, sentences=total_summary_read).replace(previously_read, '')
+        return next_summary_section, total_summary_read
 
     def search(self, query: str, lang: str = 'en') -> list([str]):
         """Search wikipedia for the given query.
@@ -132,3 +190,17 @@ class Wiki():
             self.wiki.language = lang
             is_changed = True
         return is_changed
+
+    def summarize_page(self, page: MediaWikiPage, sentences: int) -> str:
+        """Get a summary of a Wikipedia page.
+
+        The summary is also cleaned for better spoken output.
+
+        Args:
+            page: to get summary of
+            sentences: number of sentences to return
+        """
+        pymediawiki_summary = page.summarize(sentences=sentences)
+        cleaned_text = re.sub(
+            "\(.*?\)", "", pymediawiki_summary).replace('  ', ' ')
+        return cleaned_text
